Rename gdn_attention_custom_op to gdn_attention_op
